backend/requirements.txt
aniso8601==9.0.1
attrs==21.3.0
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
flask-restx==0.5.1
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.3.2
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
pluggy==1.0.0
psycopg2-binary==2.9.2
py==1.11.0
pyparsing==3.0.7
pyrsistent==0.18.0
pytest==6.2.5
pytz==2021.3
requests==2.27.1
six==1.16.0
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
